% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{adjusted_cosine_similarity}
\alias{adjusted_cosine_similarity}
\title{compute the similarity of queryProfiles and subjectProfiles}
\usage{
adjusted_cosine_similarity(
  queryProfiles,
  subjectProfiles,
  is_parallel = TRUE,
  no_cores = NULL
)
}
\arguments{
\item{queryProfiles}{a result of \code{infer_profiles_vb}}

\item{subjectProfiles}{a result of \code{infer_profiles_vb}}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores given as input}
}
\value{
a vector of Normalized adjusted cosine similarity [0-1], 
 Note: the more close to 1 indicates the profiles of region more similar,
 the more close to 0 indicates the profiles more different.
}
\description{
compute the similarity of queryProfiles and subjectProfiles by adjusted cosine 
             distance and return a vector of numeric match the queryProfiles and subjectProfiles.
}
\examples{
\dontrun{
similarity <- adjusted_cosine_similarity(queryProfiles=human_fit_profiles, 
                                         subjectProfiles=mouse_fit_profiles)
                                         
#check the number of non-NA
length(which(!is.na(similarity)))

#check the number of >0.9
length(which(similarity>0.9))

#check the number of <0.1
length(which(similarity<0.1))
}

}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
