% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_genomic_coordinate.R
\name{change_genomic_coordinate}
\alias{change_genomic_coordinate}
\title{change the genomic coordinate system of inputRegion by liftOver with chainFile}
\usage{
change_genomic_coordinate(
  inputRegion,
  chainFile,
  targetAnnotation = NULL,
  is_parallel = TRUE,
  no_cores = NULL
)
}
\arguments{
\item{inputRegion}{the region you want to change the genomic coordinate system,a \code{GRanges} object.
Generally, the inputRegin is a result of \code{cluster_sites_to_region}.}

\item{chainFile}{a chainFile of liftOver. for example, if you want mapping the the inputRegion from hg38 to mm10,
then the chainFile must be "hg38ToMm10.over.chain".
you can download from the source
\url{http://hgdownload.soe.ucsc.edu/downloads.html#liftover}}

\item{targetAnnotation}{the \code{GRanges} object to annotate the result of liftOver, a \code{GRanges} object.
Generally, the targetAnnotation is a result of \code{read_annotation}.
default value is null.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{if you have specify the is_parallel is ture ,you can specify the number of parallel cores}
}
\value{
A \code{GRanges} object corrsponding to the inputRegion in another genomic coordinate system
}
\description{
change the genomic coordinate system of inputRegion by liftOver with chainFile
}
\examples{
 
\dontrun{
filePath <- system.file("extdata", "hg38ToMm10.over.chain", package = "BSDMR")
mouse_region <- change_genomic_coordinate(human_region,filePath,mouse_anno)
}

}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
