% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sites_to_region.R
\name{cluster_sites_to_region}
\alias{cluster_sites_to_region}
\title{cluster sites to region}
\usage{
cluster_sites_to_region(
  methylation,
  annotation,
  min_sites_number = 10,
  max_distance_between_sites = 200,
  is_parallel = TRUE,
  no_cores = NULL
)
}
\arguments{
\item{methylation}{\code{GRanges} object created by \code{read_methylation_report}.}

\item{annotation}{\code{GRanges} object created by \code{read_annotation}.}

\item{min_sites_number}{the minmal number of C sites in a region. the default is 10}

\item{max_distance_between_sites}{the maxmal distance between C sites within a region.
the unit of distance is unit and the default is 200}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{if you have specify the is_parallel is ture ,you can specify the number of parallel cores}
}
\value{
A \code{GRanges} object.
}
\description{
cluster sites which meet the required suitations to region according to the annotation file
}
\examples{
\dontrun{
human_region <- cluster_sites_to_region(methylation = human_met,
                                        annotation = human_anno,is_parallel = TRUE)
}


}
\seealso{
\code{\link{read_methylation_report}},  \code{\link{read_annotation}}
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
