% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_region_object.R
\name{create_basis}
\alias{create_basis}
\alias{create_rbf_object}
\alias{basis}
\alias{create_basis_function}
\title{Create basis objects}
\usage{
create_rbf_object(
  M = 8,
  gamma = NULL,
  mus = NULL,
  eq_spaced_mus = TRUE,
  whole_region = TRUE
)
}
\arguments{
\item{M}{The number of the basis functions. In case of Fourier basis, this
number should be even, since we need to have pairs of sines and cosines and
the constant term is added by default.}

\item{gamma}{Inverse width of radial basis function.}

\item{mus}{Optional centers of the RBF.}

\item{eq_spaced_mus}{Logical, if TRUE, equally spaced centers are created,
otherwise centers are created using \code{\link[stats]{kmeans}} algorithm.}

\item{whole_region}{Logical, indicating if the centers will be evaluated
equally spaced on the whole region, or between the min and max of the
observation values.}
}
\value{
A basis object of class 'rbf'.
}
\description{
These functions create different basis objects, which can be
  used as input to complex functions in order to perform computations
  depending on the class of the basis function.
}
\examples{
human_basis_profile <- create_rbf_object(M = 8)
human_basis_mean <- create_rbf_object(M = 0)
#---------------------------------

}
\seealso{
\code{\link{design_matrix}}
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
