% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_region_object.R
\name{create_region_object}
\alias{create_region_object}
\alias{create_region_obj}
\title{Create genomic region data}
\usage{
create_region_object(
  met_dt,
  anno_dt,
  cov = 5,
  ignore_strand = TRUE,
  is_parallel = TRUE,
  no_cores = NULL
)
}
\arguments{
\item{met_dt}{A \code{GRanges} object with methylation data, whose format
should be similar to \code{\link{read_methylation_report}}.}

\item{anno_dt}{A \code{GRanges} object with annotation data, whose format
should be similar to  \code{\link{read_annotation}}.}

\item{cov}{Integer defining the minimum coverage of CpGs that each region
must contain.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{if you have specify the is_parallel is ture ,you can specify the number of parallel cores}
}
\value{
A \code{list} object containing the two elements: \itemize{ \item{
  \code{met}: A list containing methylation region data, where each entry in
  the list is an \eqn{L_{i} X D} dimensional matrix, where \eqn{L_{i}}
  denotes the number of CpGs found in region \code{i}. The columns contain
  the following information: \enumerate{ \item{ 1st column: Contains the
  locations of CpGs relative to centre. Note that the actual locations are
  scaled to the [-1,1] region. } \item{ 2nd column: Contains the total 
  number of reads at each CpG location.} \item{ 3rd column: Contains the
  methylated reads at each CpG location.} }.
  Rownames of each matrix contain the actual CpG genomic
  coordinates as <chr>:<location>. } \item{ \code{anno}: The annotation
  object.} } Note: The lengths of \code{met} and \code{anno} should match.
}
\description{
\code{create_region_object} creates genomic regions (e.g. forms
  methylation regions data) using as input methylation and annotation data
  with genomic regions of interest.
}
\examples{
\dontrun{
# Download the files and change the working directory to that location
human_met <- read_methylation_report("name_of_met_file")
human_anno <- read_annotation("name_of_anno_file")
human_obj <- create_region_object(human_met, human_anno)
}
}
\seealso{
\code{\link{read_methylation_report}}, \code{\link{read_annotation}}
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
