% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{design_matrix}
\alias{design_matrix}
\alias{designmatrix}
\alias{des_matrix}
\alias{des_mat}
\alias{design_matrix.default}
\alias{design_matrix.rbf}
\title{Generic function for creating design matrices}
\usage{
design_matrix(obj, ...)

\method{design_matrix}{default}(obj, ...)

\method{design_matrix}{rbf}(obj, obs, ...)
}
\arguments{
\item{obj}{A basis function object.}

\item{...}{Additional parameters.}

\item{obs}{A vector of observations.}
}
\value{
A design matrix object
}
\description{
These functions call the appropriate methods depending on the
  class of the object \code{obj} to create RBF design matrices.
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
