% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_fit_profiles}
\alias{human_fit_profiles}
\title{a example out of \code{infer_profiles_vb }}
\format{
An object of class \code{infer_profiles_vb_}"obs_model" with the
  following elements: \itemize{ \item{ \code{W}: An Nx(M+1) matrix with the
  optimized parameter values. Each row of the matrix corresponds to each
  element of the list X; if X is a matrix, then N = 1. The columns are of the
  same length as the parameter vector w (i.e. number of basis functions). }
  \item{ \code{W_Sigma}: A list with covariance matrices for each element row
  in W.} \item{ \code{basis}: The basis object. } \item{\code{nll_feat}: NLL
  fit feature.} \item{\code{rmse_feat}: RMSE fit feature.}
  \item{\code{coverage_feat}: CpG coverage feature.} \item{\code{lb_feat}:
  Lower Bound feature.}}
}
\usage{
human_fit_profiles
}
\value{
infer_profiles_vb
}
\description{
\code{human_basis_profile <- create_rbf_object(M = 8)
human_fit_profiles <- infer_profiles_vb(X = human_obj$met, model = "binomial",
                        basis = human_basis_profile, is_parallel = TRUE, vb_max_iter = 100)}
}
\seealso{
\code{\link{infer_profiles_vb}}
}
\keyword{datasets}
