% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_region}
\alias{human_region}
\title{a example output of \code{cluster_sites_to_region }}
\format{
A \code{GRanges} object.
The GRanges object contains three additional metadata columns:
     \itemize{ \item{ \code{id}: EnsemblID}  
     \item{ \code{center}: the center of the region,(strat+end)/2} 
     \item{ \code{annotation}:the annotation of the region, it can be promoter,downstream,gene,UTR...} }
}
\usage{
human_region
}
\value{
annotation data
}
\description{
\code{human_region <- cluster_sites_to_region(methylation = human_met,
                               annotation = human_anno,is_parallel = TRUE)}
}
\seealso{
\code{\link{cluster_sites_to_region}}
}
\keyword{datasets}
