% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_profiles.R
\name{infer_profiles_vb}
\alias{infer_profiles_vb}
\alias{infer_profile_vb}
\alias{inference_vb}
\title{Infer methylation profiles using VB}
\usage{
infer_profiles_vb(
  X,
  basis = NULL,
  H = NULL,
  w = NULL,
  alpha_0 = 0.5,
  beta_0 = 0.1,
  vb_max_iter = 100,
  epsilon_conv = 1e-05,
  is_parallel = TRUE,
  no_cores = NULL,
  is_verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{The input data, either a \code{\link[base]{matrix}} or a
\code{\link[base]{list}} of elements of length N, where each element is an
\code{L X C} matrix, where L are the total number of observations. The
first column contains the input observations x (i.e. CpG locations). If
"binomial" model then C=3, and 2nd and 3rd columns contain total number of
trials and number of successes respectively.}

\item{basis}{A 'basis' object. E.g. see \code{\link{create_rbf_object}}. If NULL,
will an RBF object will be created.}

\item{H}{Optional, design matrix of the input data X. If NULL, H will be
computed inside the function.}

\item{w}{A vector of initial parameters (i.e. coefficients of the basis
functions). If NULL, it will be initialized inside the function.}

\item{alpha_0}{Hyperparameter: shape parameter for Gamma distribution. A
Gamma distribution is used as prior for the precision parameter tau.}

\item{beta_0}{Hyperparameter: rate parameter for Gamma distribution. A Gamma
distribution is used as prior for the precision parameter tau.}

\item{vb_max_iter}{Integer denoting the maximum number of VB iterations.}

\item{epsilon_conv}{Numeric denoting the convergence threshold for VB.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_verbose}{Logical, print results during VB iterations.}

\item{...}{Additional parameters.}
}
\value{
An object of class \code{infer_profiles_vb_}"obs_model" with the
  following elements: \itemize{ \item{ \code{W}: An Nx(M+1) matrix with the
  optimized parameter values. Each row of the matrix corresponds to each
  element of the list X; if X is a matrix, then N = 1. The columns are of the
  same length as the parameter vector w (i.e. number of basis functions). }
  \item{ \code{W_Sigma}: A list with covariance matrices for each element row
  in W.} \item{ \code{basis}: The basis object. } \item{\code{nll_feat}: NLL
  fit feature.} \item{\code{rmse_feat}: RMSE fit feature.}
  \item{\code{coverage_feat}: CpG coverage feature.} \item{\code{lb_feat}:
  Lower Bound feature.}}
}
\description{
General purpose functions for inferring latent profiles for
  different observation models using Variational Bayes (VB). Current
  observation models are: 'bernoulli', 'binomial'.
}
\section{Details}{
 The modelling and mathematical details for inferring
  profiles using mean-field variational inference are explained here:
  \url{http://rpubs.com/cakapourani/} . More specifically: \itemize{\item{For
  Binomial/Bernoulli observation model check:
  \url{http://rpubs.com/cakapourani/variational-bayes-bpr}} \item{For
  Gaussian observation model check:
  \url{http://rpubs.com/cakapourani/variational-bayes-lr}}}
}

\examples{
# Example of inferring parameters for synthetic data using 8 RBFs
\dontrun{
basis_profile <- create_rbf_object(M=8)
human_fit_profiles <- infer_profiles_vb(X = human_obj$met, basis = basis_profile,
                                        is_parallel = TRUE, vb_max_iter = 100)
}

}
\seealso{
\code{\link{create_rbf_object}},  \code{\link{create_region_object}}
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
