% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_log_likelihood}
\alias{model_log_likelihood}
\alias{bpr_log_likelihood}
\alias{bpr_gradient}
\alias{betareg_log_likelihood}
\alias{betareg_gradient}
\alias{sum_weighted_bpr_lik}
\alias{sum_weighted_bpr_grad}
\alias{sum_weighted_betareg_lik}
\alias{sum_weighted_betareg_grad}
\title{model_log_likelihood}
\usage{
bpr_log_likelihood(w, X, H, lambda, is_nll)

bpr_gradient(w, X, H, lambda, is_nll)

betareg_log_likelihood(w, X, H, lambda, is_nll)

betareg_gradient(w, X, H, lambda, is_nll)

sum_weighted_bpr_lik(w, X_list, H_list, r_nk, lambda, is_nll)

sum_weighted_bpr_grad(w, X_list, H_list, r_nk, lambda, is_nll)

sum_weighted_betareg_lik(w, X_list, H_list, r_nk, lambda, is_nll)

sum_weighted_betareg_grad(w, X_list, H_list, r_nk, lambda, is_nll)
}
\arguments{
\item{w}{a vector}

\item{X}{a matrix}

\item{H}{a matrix}

\item{lambda}{a number}

\item{is_nll}{boolean}

\item{X_list}{a list}

\item{H_list}{a list}

\item{r_nk}{a vector}
}
\description{
model_log_likelihood
}
