% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_met}
\alias{mouse_met}
\title{a example output of \code{read_methylation_report}}
\format{
A \code{GRanges} object.

The GRanges object contains three additional metadata columns:
\itemize{ \item{ \code{methylated_reads}: the number if methylated reads for each C sites.}
\item{ \code{total_reads}: Total number of reads for each C sites.}
\item{ \code{type}:the C-context type: CG,CHG, or CHH}
}
}
\usage{
mouse_met
}
\value{
methylation data
}
\description{
a samll subset methylation data of mouse chr15
}
\seealso{
\code{\link{read_methylation_report}}
}
\keyword{datasets}
