% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_obj}
\alias{mouse_obj}
\title{a example output of \code{create_region_object }}
\format{
A \code{list} object containing the two elements: \itemize{ \item{
  \code{met}: A list containing methylation region data, where each entry in
  the list is an \eqn{L_{i} X D} dimensional matrix, where \eqn{L_{i}}
  denotes the number of CpGs found in region \code{i}. The columns contain
  the following information: \enumerate{ \item{ 1st column: Contains the
  locations of CpGs relative to centre. Note that the actual locations are
  scaled to the [-1,1] region. } \item{ 2nd column: Contains the total 
  number of reads at each CpG location.} \item{ 3rd column: Contains the
  methylated reads at each CpG location.} }.
  Rownames of each matrix contain the actual CpG genomic
  coordinates as <chr>:<location>. } \item{ \code{anno}: The annotation
  object.} } Note: The lengths of \code{met} and \code{anno} should match.
}
\usage{
mouse_obj
}
\value{
region_object
}
\description{
\code{mouse_obj <- create_region_object(mouse_met, mouse_region)}
}
\seealso{
\code{\link{create_region_object}}
}
\keyword{datasets}
