% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_region}
\alias{mouse_region}
\title{a example output of \code{change_genomic_coordinate }}
\format{
A \code{GRanges} object.
The GRanges object contains three additional metadata columns:
\itemize{ \item{ \code{id}: EnsemblID or liftOver FAILED,indicates the region mapping failed between seocies}
\item{ \code{center}: the center of the region,(strat+end)/2, or NA if liftOver FAILED}
\item{ \code{annotation}:the annotation of the region, it can be promoter,downstream,gene,UTR...
or the reason of liftOver FAILED.} }
}
\usage{
mouse_region
}
\value{
annotation data
}
\description{
\code{ filePath <- system.file("extdata", "hg38ToMm10.over.chain", package = "BSDMR")
                   mouse_region <- change_genomic_coordinate(human_region,filePath,mouse_anno)}
}
\seealso{
\code{\link{change_genomic_coordinate}}
}
\keyword{datasets}
