% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_infer_profiles}
\alias{plot_infer_profiles}
\title{Plot inferred methylation profiles across a region}
\usage{
plot_infer_profiles(
  region = 1,
  obj_prof,
  obj_mean = NULL,
  obs = NULL,
  title = "Inferred profiles",
  ...
)
}
\arguments{
\item{region}{Genomic region number}

\item{obj_prof}{Inferred profile, i.e. output from
\code{\link{infer_profiles_vb}}}

\item{obj_mean}{Inferred mean function, i.e. output from
\code{\link{infer_profiles_vb}}}

\item{obs}{a list result of \code{\link{create_region_object}}}

\item{title}{Plot title}

\item{...}{Additional parameters}
}
\value{
A ggplot2 object.
}
\description{
Function for plotting the inferred methylation profiles across a
given region, and optionally the mean methylation rate together with the
observed methylation data, using \code{\link{ggplot2}}.
}
\examples{
\dontrun{
# Fit methylation profiles using 8 RBFs and 0 RBF
human_basis_profile <- create_rbf_object(M = 8)
human_basis_mean <- create_rbf_object(M = 0)
human_obj <- create_region_object(human_met, human_region)
human_fit_profiles <- infer_profiles_vb(X = human_obj$met, model = "binomial",
   basis = human_basis_profile, is_parallel = TRUE, vb_max_iter = 100)
human_fit_mean <- infer_profiles_vb(X = human_obj$met, model = "binomial",
   basis = human_basis_mean, is_parallel = TRUE, vb_max_iter = 100)
p <- plot_infer_profiles(region = 44, obj_prof = human_fit_profiles,obj_mean = human_fit_mean, 
   obs = human_obj, title = paste0("Gene ID ",human_obj$anno$id[44]))
}

}
\seealso{
\code{\link{infer_profiles_vb}}, \code{\link{create_rbf_object}}, \code{\link{create_region_object}}
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
