% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_annotation}
\alias{read_annotation}
\title{Read a annotation file}
\usage{
read_annotation(file)
}
\arguments{
\item{file}{the path of annotation file}
}
\value{
A \code{GRanges} object.
The GRanges object contains three additional metadata columns:
     \itemize{ \item{ \code{id}: EnsemblID}  
     \item{ \code{center}: the center of the region,(strat+end)/2} 
     \item{ \code{annotation}:the annotation of the region, it can be promoter,downstream,gene,UTR...} }
}
\description{
\code{read_annotation} reads a file containing annotation data, which
  will be used to select genomic regions for downstream analysis.
  The annotation file format is the following: "chromosome", "EnsemblID", "name","start", "end",
  "strand".
}
\examples{
# Obtain the path to files
\dontrun{
#' file <- system.file("extdata", "human_chr22_annotation.txt", package = "BSDMR")
human_anno <- read_annotation(file)
file <- system.file("extdata", "mouse_chr15_annotation.txt", package = "BSDMR")
mouse_anno <- read_annotation(file)
}

}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
