% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_methylation_report}
\alias{read_methylation_report}
\title{Read a methylation report file}
\usage{
read_methylation_report(file, min_coverage = 4, type = "CG")
}
\arguments{
\item{file}{the path of methylation report file}

\item{min_coverage}{filter out the sites which total reads < min_coverage. the default is 4.}

\item{type}{specify the methylation context type : "CG", "CHG" or "CHH",the default is "CG".}
}
\value{
A \code{GRanges} object.
The GRanges object contains three additional metadata columns:
\itemize{ \item{ \code{methylated_reads}: the number if methylated reads for each C sites.}
\item{ \code{total_reads}: Total number of reads for each C sites.}
\item{ \code{type}:the C-context type: CG,CHG, or CHH}
}
}
\description{
\code{read_methylation_report} read a file containing methylation information using \code{readr}.
In general, the metyylation report file was generated by \code{Bismark methylation extractor}.
The methylation report file format is the following:
<chromosome> <position> <strand> <count methylated> <count unmethylated> <C-context> <trinucleotide context>
Note: the 7th column is optional.
}
\examples{
# Obtain the path to files
\dontrun{
#' file <- system.file("extdata", "example_human_chr22_CpG_report.txt", package = "BSDMR")
human_met <- read_methylation_report(file,min_coverage=4)
file <- system.file("extdata", "example_mouse_chr15_CpG_report.txt", package = "BSDMR")
mouse_met <- read_methylation_report(file,min_coverage=4)
}

}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
