% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDMR-package.R
\docType{package}
\name{BSDMR}
\alias{BSDMR}
\alias{BSDMR-package}
\title{BSDMR: a region-based method for quantitating the conservation and divergence of methylation patterns between species by high order features}
\description{
Extracting higher order methylation features by GLM to find DMRs between species.
A region-based method to find DMR.
}
\author{
Hongen Kang  \email{geneprophet@163.com}
}
\keyword{internal}
